pyproject.toml
[project]
name = "evalbot"
version = "0.1.0"
description = "教學評估表單 CLI 工具 — Claude Code 驅動"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evalbot"]
